pyproject.toml
[project]
name = "myao3"
version = "0.1.0"
description = "Event-driven autonomous bot that adapts to its environment"
requires-python = ">=3.12"
dependencies = [
    "strands-agents[otel]>=0.1,<2.0",
    "litellm>=1.0,<2.0",
    "pydantic>=2.0,<3.0",
    "pyyaml>=6.0,<7.0",
    "structlog>=24.0,<26.0",
    "aiohttp>=3.9,<4.0",
    "ulid-py>=1.1,<2.0",
    "jinja2>=3.0,<4.0",
    "python-dotenv>=1.0,<2.0",
    "ollama>=0.6.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/myao3"]

[tool.ruff]
line-length = 88
target-version = "py312"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[dependency-groups]
dev = [
    "pytest>=8.0,<10.0",
    "pytest-asyncio>=0.23,<2.0",
    "pytest-aiohttp>=1.0,<2.0",
    "pytest-cov>=4.0,<8.0",
    "ruff>=0.4,<1.0",
    "ty>=0.0.12",
]
